Tidy up commands and extend ArticleAPI.
